requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
tweepy==4.15.0
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
